Renames giphy app
