+++ packages/app-store/giphy/components/InstallAppButton.tsx
