+++ packages/app-store/giphy/package.json
