+++ packages/app-store/metadata.ts
@@ -1,7 +1,7 @@
 import { metadata as applecalendar } from "./applecalendar/_metadata";
 import { metadata as caldavcalendar } from "./caldavcalendar/_metadata";
 import { metadata as dailyvideo } from "./dailyvideo/_metadata";
-import { metadata as giphy } from "./giphyother/_metadata";
+import { metadata as giphy } from "./giphy/_metadata";
 import { metadata as googlecalendar } from "./googlecalendar/_metadata";
 import { metadata as googlevideo } from "./googlevideo/_metadata";
 import { metadata as hubspotothercalendar } from "./hubspotothercalendar/_metadata";
